API context created, queries need testing
