+++ client/src/components/LoginForm.jsx
@@ -1,14 +1,16 @@
 // see SignupForm.js for comments
 import { useState } from 'react';
 import { Form, Button, Alert } from 'react-bootstrap';
-
-import { loginUser } from '../utils/API';
+import { useAPIContext } from '../utils/API';
 import Auth from '../utils/auth';
 
 const LoginForm = () => {
   const [userFormData, setUserFormData] = useState({ email: '', password: '' });
   const [validated] = useState(false);
   const [showAlert, setShowAlert] = useState(false);
+
+  const { loginUser } = useAPIContext();
+
 
   const handleInputChange = (event) => {
     const { name, value } = event.target;
@@ -26,7 +28,7 @@ const LoginForm = () => {
     }
 
     try {
-      const response = await loginUser(userFormData);
+      const response = await loginUser({variables: {email: userFormData.email, password: userFormData.password}});
 
       if (!response.ok) {
         throw new Error('something went wrong!');
