+++ client/src/utils/Mutations.js
@@ -1,3 +1,5 @@
+import { gql } from '@apollo/client';
+
 export const CREATE_USER = gql`
   mutation createUser($username: String!, $email: String!, $password: String!) {
     createUser(username: $username, email: $email, password: $password) {
@@ -22,8 +24,9 @@ export const SAVE_BOOK = gql`
       id
       username
       savedBooks {
-        bookId: ID!
-        title: String!
+        _id
+        bookId
+        title
       }
     }
   }
@@ -35,8 +38,9 @@ export const DELETE_BOOK = gql`
       id
       username
       savedBooks {
-        bookId: ID!
-        title: String!
+        _id
+        bookId
+        title
       }
     }
   }
