+++ client/src/utils/Queries.js
@@ -1,3 +1,5 @@
+import { gql } from '@apollo/client';
+
 export const GET_USER = gql`
   query getUser($id: ID, $username: String) {
     user(id: $id, username: $username) {
